fix: track watch sessions but don't sync remote
Always track the watch session if syncing during or after playback
is enabled, but don't actually sync during the playback if syncing
during playback is not enabled. This is to make sure the syncing
after playback is not treated as an accidental event because it is
below the initial skip event count if we don't track the events
that would be fired during the playback.