fix: cache plugins exposed state
Cache the 'has plugins API' state to prevent needing to contact the server before attempting to establish a SignalR connection, among other things.

This change will fix the regression causing the auto-start SignalR option to causing Jellyfin to fail to start properly if the Shoko Server instance is available while Jellyfin is starting.