refactor: conditional merging of multiple version
Leave things as-is if everything is as intended, and only conditionally update the videos that needs to be updated. This will sped it up from 2m to 7s with a hot cache, and from 4m to 2m on a cold cache for my test server, in addition it will now NOT touch things which are already correctly linked/unlinked.