feat: add new **EXPERIMENTAL** iterative VFS generation
Added a new **EXPERIMENTAL** mode to iteratively generate the VFS for only newly imported files. There are also a few options to do a full generation every `X` generation, and to forcefully generate the whole VFS on next generation. It has no effect on generations _within_ the library, only when the library itself is generated (be it in the scheduled task or from a refresh/scan of the media folder itself).