refactor: improve description cleanup logic
Update the description cleanup logic to better accommodate the current descriptions I found on anime across AniDB. Also added logic to extract the notes to use in the next commit. With the notes extracted and with the cleanup logic enabled then it will allow the description to fall-through to the next description source (if any) if the description was only a note or otherwise cleaned-up text.