fix: use the provided list in resolver
Now that the VFS is always physical, we can use the provided file list. We previously had to iterate over the VFS folder ourselves because we might be masking the children as children of another media folder, but now that we don't need to care for that we can just as well just use the provided list.