fix: stop caring about creation time and last modified time for video file links
We've now stopped caring about the timestamps for the video file symbolic links, because from Jellyfin 10.11.1 and onwards the core doesn't use them anymore, so we don't have to support them either. The patch to set the creation date on the episodes/movies at a base-item level is still live, and hopefully is enough to mimic the older behavior we got from setting the creation time on the video file link.

As a side note on this topic, it also doesn't help that linux only have changed/modified/accessed times by default, with some file systems also having the birth time (AKA creation time). And it also doesn't help that attempting to set both the creation time and last modified time will set the _last modified time_ **twice** on linux in the core of the dotnet runtime.